Move schema update tests to org.hibernate.reactive.schema package
