+++ hibernate-reactive-core/src/test/java/org/hibernate/reactive/schema/SchemaUpdateMariaDBTestBase.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  * Copyright: Red Hat Inc. and Hibernate Authors
  */
-package org.hibernate.reactive;
+package org.hibernate.reactive.schema;
 
 import java.io.Serializable;
 import java.util.Objects;
@@ -21,6 +21,7 @@ import javax.persistence.Table;
 import javax.persistence.UniqueConstraint;
 
 import org.hibernate.cfg.Configuration;
+import org.hibernate.reactive.BaseReactiveTest;
 import org.hibernate.reactive.provider.Settings;
 import org.hibernate.reactive.testing.DatabaseSelectionRule;
 import org.hibernate.tool.schema.spi.SchemaManagementException;
@@ -32,13 +33,14 @@ import org.junit.Test;
 
 import io.vertx.ext.unit.TestContext;
 
-import static org.hibernate.reactive.containers.DatabaseConfiguration.DBType.MYSQL;
+import static org.hibernate.reactive.containers.DatabaseConfiguration.DBType.MARIA;
 import static org.hibernate.tool.schema.JdbcMetadaAccessStrategy.GROUPED;
 import static org.hibernate.tool.schema.JdbcMetadaAccessStrategy.INDIVIDUALLY;
 
-public abstract class SchemaUpdateMySqlTestBase extends BaseReactiveTest {
+public abstract class SchemaUpdateMariaDBTestBase extends BaseReactiveTest {
 
-	public static class IndividuallySchemaUpdateMySqlTestBase extends SchemaUpdateMySqlTestBase {
+	public static class IndividuallySchemaUpdateMariaDBTestBase
+			extends SchemaUpdateMariaDBTestBase {
 
 		@Override
 		protected Configuration constructConfiguration(String hbm2DdlOption) {
@@ -48,7 +50,7 @@ public abstract class SchemaUpdateMySqlTestBase extends BaseReactiveTest {
 		}
 	}
 
-	public static class GroupedSchemaUpdateMySqlTestBase extends SchemaUpdateMySqlTestBase {
+	public static class GroupedSchemaUpdateMariaDBTestBase extends SchemaUpdateMariaDBTestBase {
 
 		@Override
 		protected Configuration constructConfiguration(String hbm2DdlOption) {
@@ -66,7 +68,7 @@ public abstract class SchemaUpdateMySqlTestBase extends BaseReactiveTest {
 	}
 
 	@Rule
-	public DatabaseSelectionRule dbRule = DatabaseSelectionRule.runOnlyFor( MYSQL );
+	public DatabaseSelectionRule dbRule = DatabaseSelectionRule.runOnlyFor( MARIA );
 
 	@Before
 	@Override
@@ -121,6 +123,18 @@ public abstract class SchemaUpdateMySqlTestBase extends BaseReactiveTest {
 
 	@Test
 	public void testUpdate(TestContext context) {
+
+		// NOTE: MariaDB returns 'A' for ascending columns sorted in ascending order
+		//       for the collation column using this query:
+		//           select * from information_schema.statistics.
+		//       For some reason, it returns null for collation when explicitly
+		//       listed, as in:
+		//           select column_name, collation from information_schema.statistics
+		//       In addition, MariaDB does not support descending indexes currently
+		//       (https://jira.mariadb.org/browse/MDEV-13756)
+
+		// For now, collation will remain in the query, but checks on results
+		// returned for that column will be commented out.
 		final String indexDefinitionQuery =
 				"select column_name, collation from information_schema.statistics " +
 						"where table_schema = 'hreact' " +
@@ -184,7 +198,7 @@ public abstract class SchemaUpdateMySqlTestBase extends BaseReactiveTest {
 												.thenAccept( result -> {
 													final Object[] resultArray = (Object[]) result;
 													context.assertEquals( "id", resultArray[0] );
-													context.assertEquals( "A", resultArray[1] );
+													//context.assertEquals( "A", resultArray[1] );
 												} )
 										).thenCompose( v -> s.createNativeQuery( indexDefinitionQuery )
 												.setParameter( 1, "ASimple" )
@@ -194,9 +208,9 @@ public abstract class SchemaUpdateMySqlTestBase extends BaseReactiveTest {
 												.thenAccept( list -> {
 													context.assertEquals( 2, list.size() );
 													context.assertEquals( "aValue", ( (Object[]) list.get( 0 ) )[0] );
-													context.assertEquals( "A", ( (Object[]) list.get( 0 ) )[1] );
+													//context.assertEquals( "A", ( (Object[]) list.get( 0 ) )[1] );
 													context.assertEquals( "aStringValue", ( (Object[]) list.get( 1 ) )[0] );
-													context.assertEquals( "D", ( (Object[]) list.get( 1 ) )[1] );
+													//context.assertEquals( "D", ( (Object[]) list.get( 1 ) )[1] );
 												} )
 										).thenCompose( v -> s.createNativeQuery( indexDefinitionQuery )
 												.setParameter( 1, "ASimple" )
@@ -206,9 +220,9 @@ public abstract class SchemaUpdateMySqlTestBase extends BaseReactiveTest {
 												.thenAccept( list -> {
 													context.assertEquals( 2, list.size() );
 													context.assertEquals( "aValue", ( (Object[]) list.get( 0 ) )[0] );
-													context.assertEquals( "D", ( (Object[]) list.get( 0 ) )[1] );
+													//context.assertEquals( "D", ( (Object[]) list.get( 0 ) )[1] );
 													context.assertEquals( "data", ( (Object[]) list.get( 1 ) )[0] );
-													context.assertEquals( "A", ( (Object[]) list.get( 1 ) )[1] );
+													//context.assertEquals( "A", ( (Object[]) list.get( 1 ) )[1] );
 												} )
 										).thenCompose( v -> s.createNativeQuery( indexDefinitionQuery )
 												.setParameter( 1, "ASimple" )
@@ -218,7 +232,7 @@ public abstract class SchemaUpdateMySqlTestBase extends BaseReactiveTest {
 												.thenAccept( list -> {
 													context.assertEquals( 1, list.size() );
 													context.assertEquals( "aStringValue", ( (Object[]) list.get( 0 ) )[0] );
-													context.assertEquals( "A", ( (Object[]) list.get( 0 ) )[1] );
+													//context.assertEquals( "A", ( (Object[]) list.get( 0 ) )[1] );
 												} )
 										).thenCompose( v -> s.createNativeQuery( indexDefinitionQuery )
 												.setParameter( 1, "AOther" )
@@ -228,9 +242,9 @@ public abstract class SchemaUpdateMySqlTestBase extends BaseReactiveTest {
 												.thenAccept( list -> {
 													context.assertEquals( 2, list.size() );
 													context.assertEquals( "id1", ( (Object[]) list.get( 0 ) )[0] );
-													context.assertEquals( "A", ( (Object[]) list.get( 0 ) )[1] );
+													//context.assertEquals( "A", ( (Object[]) list.get( 0 ) )[1] );
 													context.assertEquals( "id2", ( (Object[]) list.get( 1 ) )[0] );
-													context.assertEquals( "A", ( (Object[]) list.get( 1 ) )[1] );
+													//context.assertEquals( "A", ( (Object[]) list.get( 1 ) )[1] );
 												} )
 										).thenCompose( v -> s.createNativeQuery( indexDefinitionQuery )
 												.setParameter( 1, "AAnother" )
@@ -240,7 +254,7 @@ public abstract class SchemaUpdateMySqlTestBase extends BaseReactiveTest {
 												.thenAccept( list -> {
 													context.assertEquals( 1, list.size() );
 													context.assertEquals( "id", ( (Object[]) list.get( 0 ) )[0] );
-													context.assertEquals( "A", ( (Object[]) list.get( 0 ) )[1] );
+													//context.assertEquals( "A", ( (Object[]) list.get( 0 ) )[1] );
 												} )
 										// check foreign keys
 										).thenCompose( v -> s.createNativeQuery( foreignKeyDefinitionQuery )
