+++ hibernate-reactive-core/src/test/java/org/hibernate/reactive/schema/SchemaUpdateSqlServerTestBase.java
@@ -3,10 +3,11 @@
  * SPDX-License-Identifier: Apache-2.0
  * Copyright: Red Hat Inc. and Hibernate Authors
  */
-package org.hibernate.reactive;
+package org.hibernate.reactive.schema;
 
 import java.io.Serializable;
 import java.util.Objects;
+import java.util.concurrent.CompletionStage;
 import javax.persistence.CascadeType;
 import javax.persistence.Entity;
 import javax.persistence.ForeignKey;
@@ -21,6 +22,7 @@ import javax.persistence.Table;
 import javax.persistence.UniqueConstraint;
 
 import org.hibernate.cfg.Configuration;
+import org.hibernate.reactive.BaseReactiveTest;
 import org.hibernate.reactive.provider.Settings;
 import org.hibernate.reactive.testing.DatabaseSelectionRule;
 import org.hibernate.tool.schema.spi.SchemaManagementException;
@@ -32,14 +34,19 @@ import org.junit.Test;
 
 import io.vertx.ext.unit.TestContext;
 
-import static org.hibernate.reactive.containers.DatabaseConfiguration.DBType.MARIA;
+import static org.hibernate.reactive.containers.DatabaseConfiguration.DBType.SQLSERVER;
 import static org.hibernate.tool.schema.JdbcMetadaAccessStrategy.GROUPED;
 import static org.hibernate.tool.schema.JdbcMetadaAccessStrategy.INDIVIDUALLY;
 
-public abstract class SchemaUpdateMariaDBTestBase extends BaseReactiveTest {
+public abstract class SchemaUpdateSqlServerTestBase extends BaseReactiveTest {
 
-	public static class IndividuallySchemaUpdateMariaDBTestBase
-			extends SchemaUpdateMariaDBTestBase {
+	private static final String DEFAULT_CATALOG_NAME = "master";
+
+	/**
+	 * Test INDIVIDUALLY option without setting the default catalog name
+	 */
+	public static class IndividuallySchemaUpdateSqlServerTest
+			extends SchemaUpdateSqlServerTestBase {
 
 		@Override
 		protected Configuration constructConfiguration(String hbm2DdlOption) {
@@ -49,7 +56,29 @@ public abstract class SchemaUpdateMariaDBTestBase extends BaseReactiveTest {
 		}
 	}
 
-	public static class GroupedSchemaUpdateMariaDBTestBase extends SchemaUpdateMariaDBTestBase {
+	/**
+	 * Test INDIVIDUALLY option when we set the catalog name to the default name
+	 */
+	public static class IndividuallySchemaUpdateWithCatalogTest
+			extends SchemaUpdateSqlServerTestBase {
+
+		@Override
+		protected Configuration constructConfiguration(String hbm2DdlOption) {
+			final Configuration configuration = super.constructConfiguration( hbm2DdlOption );
+			configuration.setProperty( Settings.DEFAULT_CATALOG, DEFAULT_CATALOG_NAME );
+			return configuration;
+		}
+
+		@Override
+		public String addCatalog(String name) {
+			return DEFAULT_CATALOG_NAME + "." + name;
+		}
+	}
+
+	/**
+	 * Test GROUPED option without setting the default catalog name
+	 */
+	public static class GroupedSchemaUpdateSqlServerTest extends SchemaUpdateSqlServerTestBase {
 
 		@Override
 		protected Configuration constructConfiguration(String hbm2DdlOption) {
@@ -59,15 +88,37 @@ public abstract class SchemaUpdateMariaDBTestBase extends BaseReactiveTest {
 		}
 	}
 
+	/**
+	 * Test GROUPED option when we set the catalog name to default name
+	 */
+	public static class GroupedSchemaUpdateWithCatalogNameTest extends SchemaUpdateSqlServerTestBase {
+
+		@Override
+		protected Configuration constructConfiguration(String hbm2DdlOption) {
+			final Configuration configuration = super.constructConfiguration( hbm2DdlOption );
+			configuration.setProperty( Settings.DEFAULT_CATALOG, DEFAULT_CATALOG_NAME );
+			return configuration;
+		}
+
+		@Override
+		public String addCatalog(String name) {
+			return DEFAULT_CATALOG_NAME + "." + name;
+		}
+	}
+
 	protected Configuration constructConfiguration(String hbm2DdlOption) {
 		Configuration configuration = constructConfiguration();
 		configuration.setProperty( Settings.HBM2DDL_AUTO, hbm2DdlOption );
-		configuration.setProperty( Settings.DEFAULT_CATALOG, "hreact" );
+		configuration.setProperty( Settings.DEFAULT_SCHEMA, "dbo" );
 		return configuration;
 	}
 
 	@Rule
-	public DatabaseSelectionRule dbRule = DatabaseSelectionRule.runOnlyFor( MARIA );
+	public DatabaseSelectionRule dbRule = DatabaseSelectionRule.runOnlyFor( SQLSERVER );
+
+	public String addCatalog(String name) {
+		return name;
+	}
 
 	@Before
 	@Override
@@ -76,8 +127,25 @@ public abstract class SchemaUpdateMariaDBTestBase extends BaseReactiveTest {
 		createHbm2ddlConf.addAnnotatedClass( ASimpleFirst.class );
 		createHbm2ddlConf.addAnnotatedClass( AOther.class );
 
-		test( context, setupSessionFactory( createHbm2ddlConf )
-				.thenCompose( v -> factoryManager.stop() ) );
+		test( context, dropSequenceIfExists( createHbm2ddlConf )
+				.thenCompose( ignore -> setupSessionFactory( createHbm2ddlConf )
+				.thenCompose( v -> factoryManager.stop() ) ) );
+	}
+
+	// See HHH-14835: Vert.x throws an exception when the catalog is specified.
+	// Because it happens during schema creation, the error is ignored and the build won't fail
+	// if one of the previous tests has already created the sequence.
+	// This method makes sure that the sequence is deleted if it exists, so that these tests
+	// fail consistently when the wrong ORM version is used.
+	private CompletionStage<Void> dropSequenceIfExists(Configuration createHbm2ddlConf) {
+		return setupSessionFactory( createHbm2ddlConf )
+				.thenCompose( v -> getSessionFactory()
+						.withTransaction( (session, transaction) -> session
+								// No need to add the catalog name because MSSQL doesn't support it
+								.createNativeQuery( "drop sequence if exists dbo.hibernate_sequence" )
+								.executeUpdate() ) )
+				.handle( (res, err) -> null )
+				.thenCompose( v -> factoryManager.stop() );
 	}
 
 	@After
@@ -93,61 +161,49 @@ public abstract class SchemaUpdateMariaDBTestBase extends BaseReactiveTest {
 				.thenCompose( v -> factoryManager.stop() ) );
 	}
 
-
 	@Test
 	public void testValidationSucceed(TestContext context) {
-		Configuration createHbm2ddlConf = constructConfiguration( "validate" );
-		createHbm2ddlConf.addAnnotatedClass( ASimpleFirst.class );
-		createHbm2ddlConf.addAnnotatedClass( AOther.class );
+		Configuration configuration = constructConfiguration( "validate" );
+		configuration.addAnnotatedClass( ASimpleFirst.class );
+		configuration.addAnnotatedClass( AOther.class );
 
-		test( context, setupSessionFactory( createHbm2ddlConf ) );
+		test( context, setupSessionFactory( configuration ) );
 	}
 
-	//TODO: I'm just checking that the validation fails because the table is missing, but we need more tests to check that
-	//      it fails for other scenarios: missing column, wrong type (?) and so on. (I don't know exactly what cases `validate`
-	//      actually checks).
+	//TODO: We need more tests to check that it fails for other scenarios:
+	//      missing column, wrong type (?) and so on. (I don't know exactly what cases `validate` actually checks).
 	@Test
 	public void testValidationFails(TestContext context) {
-		Configuration createHbm2ddlConf = constructConfiguration( "validate" );
-		createHbm2ddlConf.addAnnotatedClass( AAnother.class );
+		Configuration configuration = constructConfiguration( "validate" );
+		configuration.addAnnotatedClass( AAnother.class );
 
-		test( context, setupSessionFactory( createHbm2ddlConf )
+		final String errorMessage = "Schema-validation: missing table [" + addCatalog( "dbo.AAnother" ) + "]";
+		test( context, setupSessionFactory( configuration )
 				.handle( (unused, throwable) -> {
 					context.assertNotNull( throwable );
 					context.assertEquals( throwable.getClass(), SchemaManagementException.class );
-					context.assertEquals( throwable.getMessage(), "Schema-validation: missing table [hreact.AAnother]" );
+					context.assertEquals( throwable.getMessage(), errorMessage );
 					return null;
 				} ) );
 	}
 
 	@Test
 	public void testUpdate(TestContext context) {
-
-		// NOTE: MariaDB returns 'A' for ascending columns sorted in ascending order
-		//       for the collation column using this query:
-		//           select * from information_schema.statistics.
-		//       For some reason, it returns null for collation when explicitly
-		//       listed, as in:
-		//           select column_name, collation from information_schema.statistics
-		//       In addition, MariaDB does not support descending indexes currently
-		//       (https://jira.mariadb.org/browse/MDEV-13756)
-
-		// For now, collation will remain in the query, but checks on results
-		// returned for that column will be commented out.
 		final String indexDefinitionQuery =
-				"select column_name, collation from information_schema.statistics " +
-						"where table_schema = 'hreact' " +
-						"and table_name = ? and index_name = ? and non_unique = ? " +
-						"order by seq_in_index";
+				"select COL_NAME(ic.object_id, ic.column_id), ic.is_descending_key " +
+					"from sys.indexes i inner join sys.index_columns ic " +
+					"on ic.object_id = i.object_id and ic.index_id = i.index_id " +
+					"where OBJECT_NAME(i.object_id) = ? and i.name = ? and i.is_unique = ? " +
+					"order by ic.key_ordinal";
 
 		final String foreignKeyDefinitionQuery =
-				"select column_name, referenced_column_name " +
-						"from information_schema.key_column_usage " +
-						"where table_schema = 'hreact' and " +
-						"constraint_name = ? and " +
-						"table_name = ? and " +
-						"referenced_table_name = ? " +
-						"order by ordinal_position";
+				"select COL_NAME( parent_object_id, parent_column_id ) as col_name, " +
+						"COL_NAME( referenced_object_id, referenced_column_id) as ref_col_name " +
+						"from sys.foreign_key_columns " +
+						"where OBJECT_NAME( constraint_object_id ) = ? " +
+						"and OBJECT_NAME( parent_object_id ) = ? " +
+						"and OBJECT_NAME( referenced_object_id ) = ? " +
+						"order by constraint_column_id";
 
 		final ASimpleNext aSimple = new ASimpleNext();
 		aSimple.aValue = 9;
@@ -191,72 +247,42 @@ public abstract class SchemaUpdateMariaDBTestBase extends BaseReactiveTest {
 										} )
 										.thenCompose( v -> s.createNativeQuery( indexDefinitionQuery )
 												.setParameter( 1, "ASimple" )
-												.setParameter( 2, "PRIMARY" )
-												.setParameter( 3, 0 )
-												.getSingleResult()
-												.thenAccept( result -> {
-													final Object[] resultArray = (Object[]) result;
-													context.assertEquals( "id", resultArray[0] );
-													//context.assertEquals( "A", resultArray[1] );
-												} )
-										).thenCompose( v -> s.createNativeQuery( indexDefinitionQuery )
-												.setParameter( 1, "ASimple" )
 												.setParameter( 2, "i_asimple_avalue_astringValue" )
-												.setParameter( 3, 1 )
+												.setParameter( 3, 0 )
 												.getResultList()
 												.thenAccept( list -> {
 													context.assertEquals( 2, list.size() );
-													context.assertEquals( "aValue", ( (Object[]) list.get( 0 ) )[0] );
-													//context.assertEquals( "A", ( (Object[]) list.get( 0 ) )[1] );
-													context.assertEquals( "aStringValue", ( (Object[]) list.get( 1 ) )[0] );
-													//context.assertEquals( "D", ( (Object[]) list.get( 1 ) )[1] );
+													context.assertEquals( "aValue", ( (Object[]) list.get(0) )[0] );
+													context.assertEquals( false, ( (Object[]) list.get(0) )[1] );
+													context.assertEquals( "aStringValue", ( (Object[]) list.get(1) )[0] );
+													context.assertEquals( true, ( (Object[]) list.get(1) )[1] );
 												} )
-										).thenCompose( v -> s.createNativeQuery( indexDefinitionQuery )
+										)
+										.thenCompose( v -> s.createNativeQuery( indexDefinitionQuery )
 												.setParameter( 1, "ASimple" )
 												.setParameter( 2, "i_asimple_avalue_data" )
+												.setParameter( 3, 0 )
+												.getResultList()
+												.thenAccept( list -> {
+													context.assertEquals( 2, list.size() );
+													context.assertEquals( "aValue", ( (Object[]) list.get(0) )[0] );
+													context.assertEquals( true, ( (Object[]) list.get(0) )[1] );
+													context.assertEquals( "data", ( (Object[]) list.get(1) )[0] );
+													context.assertEquals( false, ( (Object[]) list.get(1) )[1] );
+												} )
+										)
+										.thenCompose( v -> s.createNativeQuery( indexDefinitionQuery )
+												.setParameter( 1, "ASimple" )
+												.setParameter( 2, "u_asimple_astringvalue" )
 												.setParameter( 3, 1 )
 												.getResultList()
 												.thenAccept( list -> {
-													context.assertEquals( 2, list.size() );
-													context.assertEquals( "aValue", ( (Object[]) list.get( 0 ) )[0] );
-													//context.assertEquals( "D", ( (Object[]) list.get( 0 ) )[1] );
-													context.assertEquals( "data", ( (Object[]) list.get( 1 ) )[0] );
-													//context.assertEquals( "A", ( (Object[]) list.get( 1 ) )[1] );
-												} )
-										).thenCompose( v -> s.createNativeQuery( indexDefinitionQuery )
-												.setParameter( 1, "ASimple" )
-												.setParameter( 2, "u_asimple_astringvalue" )
-												.setParameter( 3, 0 )
-												.getResultList()
-												.thenAccept( list -> {
 													context.assertEquals( 1, list.size() );
-													context.assertEquals( "aStringValue", ( (Object[]) list.get( 0 ) )[0] );
-													//context.assertEquals( "A", ( (Object[]) list.get( 0 ) )[1] );
+													context.assertEquals( "aStringValue", ( (Object[]) list.get(0) )[0] );
+													context.assertEquals( false, ( (Object[]) list.get(0) )[1] );
 												} )
-										).thenCompose( v -> s.createNativeQuery( indexDefinitionQuery )
-												.setParameter( 1, "AOther" )
-												.setParameter( 2, "PRIMARY" )
-												.setParameter( 3, 0 )
-												.getResultList()
-												.thenAccept( list -> {
-													context.assertEquals( 2, list.size() );
-													context.assertEquals( "id1", ( (Object[]) list.get( 0 ) )[0] );
-													//context.assertEquals( "A", ( (Object[]) list.get( 0 ) )[1] );
-													context.assertEquals( "id2", ( (Object[]) list.get( 1 ) )[0] );
-													//context.assertEquals( "A", ( (Object[]) list.get( 1 ) )[1] );
-												} )
-										).thenCompose( v -> s.createNativeQuery( indexDefinitionQuery )
-												.setParameter( 1, "AAnother" )
-												.setParameter( 2, "PRIMARY" )
-												.setParameter( 3, 0 )
-												.getResultList()
-												.thenAccept( list -> {
-													context.assertEquals( 1, list.size() );
-													context.assertEquals( "id", ( (Object[]) list.get( 0 ) )[0] );
-													//context.assertEquals( "A", ( (Object[]) list.get( 0 ) )[1] );
-												} )
-										// check foreign keys
-										).thenCompose( v -> s.createNativeQuery( foreignKeyDefinitionQuery )
+										)
+										.thenCompose( v -> s.createNativeQuery( foreignKeyDefinitionQuery )
 												.setParameter( 1, "fk_asimple_aother" )
 												.setParameter( 2, "ASimple" )
 												.setParameter( 3, "AOther" )
@@ -279,7 +305,7 @@ public abstract class SchemaUpdateMariaDBTestBase extends BaseReactiveTest {
 													context.assertEquals( "id", ( (Object[]) result )[1] );
 												} )
 										)
-						) )
+								) )
 		);
 
 	}
